// bridge/typescript/src/_generated/dir/tree/key.ts
export type Name =
    /** A regular identifier. */
    | {
          readonly kind: "identifier";
          readonly identifier: StringId;
      }
    /** A string identifier. */
    | {
          readonly kind: "string";
          readonly string: StringId;
      }
    /** A positional index. */
    | {
          readonly kind: "index";
          readonly index: number;
      }
;

export const Name = {
    ...NameImpl,

    /** A regular identifier. */
    identifier(identifier: StringId): Name {
        return { kind: "identifier", identifier };
    },

    /** A string identifier. */
    "string"(string_: StringId): Name {
        return { kind: "string", string: string_ };
    },

    /** A positional index. */
    index(index: number): Name {
        return { kind: "index", index };
    },

    /** Encode this value. */
    encode(writer: BinaryWriter, value: Name): void {
        encodeName(writer, value);
    },

    /** Decode one Name. */
    decode(reader: BinaryReader): Name {
        return decodeName(reader);
    },

    /** Return this value as JSON. */
    toJson(value: Name): Json {
        return toJsonName(value);
    },

    /** Return one Name from one JSON value. */
    fromJson(value: Json): Name {
        return fromJsonName(value);
    },
};
// ...
export type Key =
    /** A named key. */
    | {
          readonly kind: "name";
          readonly name: Name;
      }
    /** A private key. */
    | {
          readonly kind: "private";
          readonly private: StringId;
      }
    /** A dynamic value-space key. */
    | {
          readonly kind: "expression";
          readonly expression: LocalNodeId;
      }
;

export const Key = {
    ...KeyImpl,

    /** A named key. */
    name(name: Name): Key {
        return { kind: "name", name };
    },

    /** A private key. */
    "private"(private_: StringId): Key {
        return { kind: "private", private: private_ };
    },

    /** A dynamic value-space key. */
    expression(expression: LocalNodeId): Key {
        return { kind: "expression", expression };
    },

    /** Encode this value. */
    encode(writer: BinaryWriter, value: Key): void {
        encodeKey(writer, value);
    },

    /** Decode one Key. */
    decode(reader: BinaryReader): Key {
        return decodeKey(reader);
    },

    /** Return this value as JSON. */
    toJson(value: Key): Json {
        return toJsonKey(value);
    },

    /** Return one Key from one JSON value. */
    fromJson(value: Json): Key {
        return fromJsonKey(value);
    },
};